requirements.txt
torch
gdown
google-cloud-storage
torchvision
argparse
numpy
Pillow
absl-py
tqdm
matplotlib
futures
git+git://github.com/saurabhgarg1996/robustness
git+git://github.com/saurabhgarg1996/calibration
wilds
crcmod
scikit-learn
seaborn
latex
IProgress
ipywidgets
networkx
transformers>=3.5.0